# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Licitações",
        "description": "Consulta e análise histórica de licitações públicas federais.\n\nSomente leitura. Os dados vêm do Portal da Transparência, competências\n201301 a 202404 - a fonte foi descontinuada com a transição para a Lei\n14.133/2021.\n\nAs análises leem tabelas materializadas por job em lote; nenhum endpoint\ncalcula agregação sobre as tabelas de volume.\n\nO recorte de período usa `competencia` (AAAAMM) e não a data de abertura,\nporque 72,6% das datas de abertura são nulas na série.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Base da API"
        }
    ],
    "paths": {
        "/analytics/evolucao": {
            "get": {
                "tags": [
                    "Análises"
                ],
                "summary": "Série temporal de quantidade e valor por competência",
                "description": "Evolução temporal. Medido em 19 ms sobre serie_mensal.",
                "operationId": "8da154758ad2d9e711f5a4a37ae7ddd5",
                "parameters": [
                    {
                        "name": "competencia_de",
                        "in": "query",
                        "description": "AAAAMM",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "competencia_ate",
                        "in": "query",
                        "description": "AAAAMM",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Série ordenada por competência",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PontoDaSerie"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "competencias": {
                                                    "type": "integer"
                                                },
                                                "competencias_parciais": {
                                                    "description": "Motivo de cada competência marcada como parcial.",
                                                    "type": "object",
                                                    "additionalProperties": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/analytics/modalidades": {
            "get": {
                "tags": [
                    "Análises"
                ],
                "summary": "Distribuição por modalidade",
                "description": "Distribuição por modalidade. Medido em 26 ms.",
                "operationId": "ab330a6d3e7f5e969f6decc60d9bbaf7",
                "parameters": [
                    {
                        "name": "competencia_de",
                        "in": "query",
                        "description": "AAAAMM",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "competencia_ate",
                        "in": "query",
                        "description": "AAAAMM",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Modalidades ordenadas por quantidade"
                    }
                }
            }
        },
        "/analytics/orgaos": {
            "get": {
                "tags": [
                    "Análises"
                ],
                "summary": "Ranking de órgãos por valor",
                "description": "Ranking de órgãos. Lê serie_mensal.",
                "operationId": "8db05367d66c06991e0c616e4ff6f4cb",
                "parameters": [
                    {
                        "name": "competencia_de",
                        "in": "query",
                        "description": "AAAAMM",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "competencia_ate",
                        "in": "query",
                        "description": "AAAAMM",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Órgãos ordenados por valor total"
                    }
                }
            }
        },
        "/analytics/fornecedores": {
            "get": {
                "tags": [
                    "Análises"
                ],
                "summary": "Ranking de fornecedores por valor",
                "description": "Sem filtro de período lê a tabela global; com filtro, a tabela por competência. A granularidade usada vem em meta.",
                "operationId": "74bbca41108454d7e841b4ee8507421a",
                "parameters": [
                    {
                        "name": "competencia_de",
                        "in": "query",
                        "description": "AAAAMM",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "competencia_ate",
                        "in": "query",
                        "description": "AAAAMM",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 100
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fornecedores ordenados por valor total",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RankingFornecedor"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "granularidade": {
                                                    "type": "string",
                                                    "enum": [
                                                        "global",
                                                        "por_competencia"
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PontoDaSerie": {
                "properties": {
                    "competencia": {
                        "type": "string",
                        "example": "202401"
                    },
                    "quantidade_licitacoes": {
                        "type": "integer"
                    },
                    "valor_total": {
                        "type": "string",
                        "nullable": true
                    },
                    "parcial": {
                        "description": "true quando a competência é publicada incompleta pela fonte. O motivo vem em meta.competencias_parciais.",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "RankingFornecedor": {
                "properties": {
                    "cnpj": {
                        "type": "string"
                    },
                    "nome": {
                        "type": "string",
                        "nullable": true
                    },
                    "itens_vencidos": {
                        "type": "integer"
                    },
                    "licitacoes_distintas": {
                        "type": "integer"
                    },
                    "valor_total": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
...
}